Don't bind to cluster-admin role if defaultNamespace is defined
minibroker is configured with a defaultNamespace when being used as a
Cloud Foundry service broker because CF does not pass along a namespace
to the broker.  That means all services will be instantiated inside the
defaultNamespace.

There are 2 configuration choices: the defaultNamespace is identical to
the release namespace, or it is separate.

This patch will create the defaultNamespace (if specified) and create a
role providing unlimited access to it, which will be bound to the minibroker
service account in the release namespace.

It will also bind a role granting all operations on configmaps inside the
release namespace to the minibroker service account because minibroker
creates and updates a configmap to track configuration information.

In case the defaultNamespace is identical to the release namespace, the
namespace is already created by helm. Also the configmap role is then
redundant with the unlimited access provided to the defaultNamespace and
therefore omitted.

If the defaultNamespace is not defined, the chart will continue to create
a cluster role binding to the cluster-admin role, just as before.